requirements.txt

fastapi
uvicorn
streamlit
faiss-cpu
pydantic
pymongo
openai>=1.0.0
python-dotenv
nltk
